Finding minimum element of the pre-declared array
Code snippet iterates over an array, compares each element with a value stored in R0, and updates R0 with the current element if it is less. The loop continues until all elements of the array have been processed. Each element takes 4 bytes, so to move from arr[0] to arr[1] we need LDR with a post-increment of 4 bytes.