Simple factorial calculator
Code uses nested loops to calculate the factorial of a given number. The inner loop is based on the multiply.asm from this repository. The outer loop (while) controls the calculation until the counter exceeds the given number, and the inner loop (for) gathers the sum of the numbers from 2 to the current counter value.